Move the tests from test_refine_rb to test_bundles.
Also, removes an undocumented two-liner and fix a bug that
happens when an indexing array is empty.
